cleaned image_tools.py
