+++ pyTEMlib/image_tools.py
@@ -21,9 +21,8 @@ import struct
 
 import sidpy
 from .file_tools import *
-
+from .probe_tools import *
 import sys
-
 
 import itertools
 from itertools import product
@@ -52,7 +51,6 @@ from skimage.feature import blob_log  # blob_dog, blob_doh
 
 from sklearn.feature_extraction import image
 from sklearn.utils.extmath import randomized_svd
-from sklearn.cluster import KMeans
 
 _SimpleITK_present = True
 try:
@@ -663,105 +661,6 @@ def histogram_plot(image_tags):
     return span
 
 
-def find_atoms(im, tags, verbose=False):
-    
-    if 'rel_blob_size' not in tags:
-        tags['rel_blob_size'] = .4  # between 0 and 1 nromally around 0.5
-        tags['source_size'] = 0.06  # in nm gives the size of the atoms or resolution
-        tags['nearest_neighbours'] = 7  # up to this number nearest neighbours are evaluated (normally 7)
-        tags['threshold'] = 0.15  # between 0.01 and 0.1
-        tags['rim_size'] = 2  # size of rim in multiples of source size
-        
-    rel_blob_size = tags['rel_blob_size']  # between 0 and 1 nromaly around 0.5
-    source_size = tags['source_size']  # in nm gives the size of the atoms
-    nearest_neighbours = tags['nearest_neighbours']  # up to this number nearest neighbours are evaluated (normally 7)
-    threshold = tags['threshold']  # between 0.01 and 0.1 
-    rim_size = tags['rim_size']  # sizeof rim in multiples of resolution
-    pixel_size = tags['pixel_size']
-                      
-    rim_width = rim_size*source_size/pixel_size
-    
-    # Get a noise free image: reduced
-    # pixel_size = FOV/im.shape[0]
-    reduced_image = clean_svd(im, pixel_size=pixel_size, source_size=source_size)
-
-    reduced_image = reduced_image-reduced_image.min()
-    reduced_image = reduced_image/reduced_image.max()
-
-    tags['reduced_image'] = reduced_image
-    patch_size = im.shape[0]-reduced_image.shape[0]
-    tags['patch_size'] = patch_size
-    print(f' Use {patch_size} x {patch_size} pixels for image-patch of atoms')
-
-    # Find atoms    
-    thresh = reduced_image.std()*threshold
-    blobs = blob_log(np.array(reduced_image), max_sigma=source_size/pixel_size, threshold=thresh)
-    
-    atoms = []
-    
-    for blob in blobs:
-        y, x, r = blob
-        if r > patch_size*rel_blob_size:
-            atoms.append([x+patch_size/2, y+patch_size/2, r])
-
-    atoms = np.array(atoms)
-    # Determine Rim atoms
-    rim_left = np.where(atoms[:, 0] < rim_width)[0].flatten()
-    rim_right = np.where(atoms[:, 0] > im.shape[0]-rim_width)[0].flatten()
-    rim_up = np.where(atoms[:, 1] < rim_width)[0].flatten()
-    rim_down = np.where(atoms[:, 1] > im.shape[1]-rim_width)[0].flatten()
-
-    rim_indices = np.concatenate((rim_left, rim_right, rim_up, rim_down))
-
-    rim_atoms = list(np.unique(rim_indices))
-    mid_atoms_list = np.setdiff1d(np.arange(len(atoms)), rim_atoms)
-    
-    mid_atoms = np.array(atoms)[mid_atoms_list]
-    if verbose:
-        print(f'Evaluated {len(mid_atoms)} atom positions, out of {len(atoms)} atoms')
-    tags['atoms'] = atoms
-    tags['mid_atoms'] = mid_atoms
-    tags['rim_atoms'] = rim_atoms
-    tags['number_of_atoms'] = len(atoms)
-    tags['number_of_evaluated_atoms'] = len(mid_atoms)
-    return tags
-
-
-def atoms_clustering(atoms, mid_atoms, number_of_clusters=3, nearest_neighbours=7):
-    # get distances
-    nn_tree = cKDTree(np.array(atoms)[:, 0:2])
-
-    distances, indices = nn_tree.query(np.array(mid_atoms)[:, 0:2], nearest_neighbours)
-
-    # CLustering
-    k_means = KMeans(n_clusters=number_of_clusters, random_state=0)  # Fixing the RNG in kmeans
-    k_means.fit(distances)
-    clusters = k_means.predict(distances)
-    return clusters, distances, indices
-
-
-def voronoi(atoms, tags):
-    im = tags['image']
-    vor = Voronoi(np.array(atoms)[:, 0:2])  # Plot it:
-    rim_vertices = []
-    for i in range(len(vor.vertices)):
-        if (vor.vertices[i, 0:2] < 0).any() or (vor.vertices[i, 0:2] > im.shape[0]-5).any():
-            rim_vertices.append(i)
-    rim_vertices = set(rim_vertices)
-    mid_vertices = list(set(np.arange(len(vor.vertices))).difference(rim_vertices))
-
-    mid_regions = []
-    for region in vor.regions:  # Check all Voronoi polygons
-        if all(x in mid_vertices for x in region) and len(region) > 1:  # we get a lot of rim (-1) and empty and regions
-            mid_regions.append(region)
-    tags['atoms']['voronoi'] = vor
-    tags['atoms']['voronoi_vertices'] = vor.vertices
-    tags['atoms']['voronoi_regions'] = vor.regions
-    tags['atoms']['voronoi_midVerticesIndices'] = mid_vertices
-    tags['atoms']['voronoi_midVertices'] = vor.vertices[mid_vertices]
-    tags['atoms']['voronoi_midRegions'] = mid_regions
-
-
 def clean_svd(im, pixel_size=1, source_size=5):
     patch_size = int(source_size/pixel_size)
     if patch_size < 3:
@@ -934,200 +833,6 @@ def align_crystal_reflections(spots, crystals):
         
     return crystal_reflections_polar, angles
 
-
-def plot_image(tags):
-    if 'axis' in tags:
-        pixel_size = tags['axis']['0']['scale']
-        units = tags['axis']['0']['units']
-    elif 'pixel_size' not in tags:
-        pixel_size = 1
-        units = 'px'
-    else:
-        pixel_size = tags['pixel_size']
-        units = 'nm'
-
-    image = tags['data'].T
-    FOV = pixel_size*image.shape[0]
-    plt.imshow(image, cmap='gray', extent=(0, FOV, 0, FOV))
-    if 'basename' in tags:
-        plt.title(tags['basename'])
-
-    plt.show()
-
-
-# ## Aberration Function for Probe calculations
-def make_chi1(phi, theta, wl, ab, c1_include):
-    """
-    ###
-    # Aberration function chi without defocus
-    ###
-    """
-    t0 = np.power(theta, 1) / 1 * (float(ab['C01a']) * np.cos(1 * phi) + float(ab['C01b']) * np.sin(1 * phi))
-
-    if c1_include == 1:  # First and second terms
-        t1 = np.power(theta, 2) / 2 * (ab['C10'] + ab['C12a'] * np.cos(2 * phi) + ab['C12b'] * np.sin(2 * phi))
-    elif c1_include == 2:  # Second terms only
-        t1 = np.power(theta, 2) / 2 * (ab['C12a'] * np.cos(2 * phi) + ab['C12b'] * np.sin(2 * phi))
-    else:  # none for zero
-        t1 = t0 * 0.
-    t2 = np.power(theta, 3) / 3 * (ab['C21a'] * np.cos(1 * phi) + ab['C21b'] * np.sin(1 * phi)
-                                   + ab['C23a'] * np.cos(3 * phi) + ab['C23b'] * np.sin(3 * phi))
-
-    t3 = np.power(theta, 4) / 4 * (ab['C30']
-                                   + ab['C32a'] * np.cos(2 * phi)
-                                   + ab['C32b'] * np.sin(2 * phi)
-                                   + ab['C34a'] * np.cos(4 * phi)
-                                   + ab['C34b'] * np.sin(4 * phi))
-
-    t4 = np.power(theta, 5) / 5 * (ab['C41a'] * np.cos(1 * phi)
-                                   + ab['C41b'] * np.sin(1 * phi)
-                                   + ab['C43a'] * np.cos(3 * phi)
-                                   + ab['C43b'] * np.sin(3 * phi)
-                                   + ab['C45a'] * np.cos(5 * phi)
-                                   + ab['C45b'] * np.sin(5 * phi))
-
-    t5 = np.power(theta, 6) / 6 * (ab['C50']
-                                   + ab['C52a'] * np.cos(2 * phi)
-                                   + ab['C52b'] * np.sin(2 * phi)
-                                   + ab['C54a'] * np.cos(4 * phi)
-                                   + ab['C54b'] * np.sin(4 * phi)
-                                   + ab['C56a'] * np.cos(6 * phi)
-                                   + ab['C56b'] * np.sin(6 * phi))
-
-    chi = t0 + t1 + t2 + t3 + t4 + t5
-    if 'C70' in ab:
-        chi += np.power(theta, 8) / 8 * (ab['C70'])
-    return chi * 2 * np.pi / wl  # np.power(theta,6)/6*(  ab['C50'] )
-
-
-def probe2(ab, size_x, size_y, tags, verbose=False):
-    """
-    **********************************************
-    * This function creates a incident STEM probe
-    * at position (0,0)
-    * with parameters given in ab dictionary
-    *
-    * The following Aberration functions are being used:
-    * 1) ddf = Cc*de/E  but not  + Cc2*(de/E)^2,
-    *    Cc, Cc2 = chrom. Aber. (1st, 2nd order) [1]
-    * 2) chi(qx,qy) = (2*pi/lambda)*{0.5*C1*(qx^2+qy^2)+
-    *                 0.5*C12a*(qx^2-qy^2)+
-    *                 C12b*qx*qy+
-    *                 C21a/3*qx*(qx^2+qy^2)+
-    *                 ...
-    *                 +0.5*C3*(qx^2+qy^2)^2
-    *                 +0.125*C5*(qx^2+qy^2)^3
-    *                 ... (need to finish)
-    *
-    *
-    *    qx = acos(kx/K), qy = acos(ky/K)
-    *
-    * References:
-    * [1] J. Zach, M. Haider,
-    *    "Correction of spherical and Chromatic Aberration
-    *     in a low Voltage SEM", Optik 98 (3), 112-118 (1995)
-    * [2] O.L. Krivanek, N. Delby, A.R. Lupini,
-    *    "Towards sub-Angstroem Electron Beams",
-    *    Ultramicroscopy 78, 1-11 (1999)
-    *
-    *********************************************'''
-    ####
-    # Internally reciprocal lattice vectors in 1/nm or rad.
-    # All calculations of chi in angles.
-    # All aberration coefficients in nm
-    """
-
-    if 'fov' not in ab:
-        if 'fov' not in tags:
-            print(' need field of view in tags ')
-        else:
-            ab['fov'] = tags['fov']
-
-    if 'convAngle' not in ab:
-        ab['convAngle'] = 30  # in mrad
-
-    ap_angle = ab['convAngle'] / 1000.0  # in rad
-
-    e0 = ab['EHT'] = float(ab['EHT'])  # acceleration voltage in ev
-
-    # defocus = ab['C10']
-
-    if 'C01a' not in ab:
-        ab['C01a'] = 0.
-    if 'C01b' not in ab:
-        ab['C01b'] = 0.
-
-    if 'C50' not in ab:
-        ab['C50'] = 0.
-    if 'C70' not in ab:
-        ab['C70'] = 0.
-
-    if 'Cc' not in ab:
-        ab['Cc'] = 1.3e6  # Cc in  nm
-
-    def get_wl():
-        h = 6.626 * 10 ** -34
-        m0 = 9.109 * 10 ** -31
-        ev = 1.602 * 10 ** -19 * e0
-        c = 2.998 * 10 ** 8
-        return h / np.sqrt(2 * m0 * ev * (1 + ev / (2 * m0 * c ** 2))) * 10 ** 9
-
-    wl = get_wl()
-    if verbose:
-        print('Acceleration voltage {0:}kV => wavelength {1:.2f}pm'.format(int(e0 / 1000), wl * 1000))
-    ab['wavelength'] = wl
-
-    # Reciprocal plane in 1/nm
-    dk = 1 / ab['fov']
-    kx = np.array(dk * (-size_x / 2. + np.arange(size_x)))
-    ky = np.array(dk * (-size_y / 2. + np.arange(size_y)))
-    t_xv, t_yv = np.meshgrid(kx, ky)
-
-    # define reciprocal plane in angles
-    phi = np.arctan2(t_xv, t_yv)
-    theta = np.arctan2(np.sqrt(t_xv ** 2 + t_yv ** 2), 1 / wl)
-
-    # calculate chi but omit defocus
-    chi = np.fft.ifftshift(make_chi1(phi, theta, wl, ab, 2))
-    probe = np.zeros((size_x, size_y))
-
-    # Aperture function
-    mask = theta >= ap_angle
-
-    # Calculate probe with Cc
-
-    for i in range(len(ab['zeroLoss'])):
-        df = ab['C10'] + ab['Cc'] * ab['zeroEnergy'][i] / e0
-        if verbose:
-            print('defocus due to Cc: {0:.2f} nm with weight {1:.2f}'.format(df, ab['zeroLoss'][i]))
-        # Add defocus
-        chi2 = chi + np.power(theta, 2) / 2 * df
-        # Calculate exponent of - i * chi
-        chi_t = np.fft.ifftshift(np.vectorize(complex)(np.cos(chi2), -np.sin(chi2)))
-        # Apply aperture function
-        chi_t[mask] = 0.
-        # inverse fft of aberration function
-        i2 = np.fft.fftshift(np.fft.ifft2(np.fft.ifftshift(chi_t)))
-        # add intensities
-        probe = probe + np.real(i2 * np.conjugate(i2)).T * ab['zeroLoss'][i]
-
-    ab0 = {}
-    for key in ab:
-        ab0[key] = 0.
-    # chiIA = np.fft.fftshift(make_chi1(phi, theta, wl, ab0, 0))  # np.ones(chi2.shape)*2*np.pi/wl
-    chi_i = np.ones((size_y, size_x))
-    chi_i[mask] = 0.
-    i2 = np.fft.fftshift(np.fft.ifft2(np.fft.ifftshift(chi_i)))
-    ideal = np.real(i2 * np.conjugate(i2))
-
-    probe_f = np.fft.fft2(probe, probe.shape) + 1e-12
-    ideal_f = np.fft.fft2(ideal, probe.shape)
-    fourier_space_division = ideal_f / probe_f
-    probe_r = (np.fft.ifft2(fourier_space_division, probe.shape))
-
-    return probe / sum(ab['zeroLoss']), np.real(probe_r)
-
-
 # Deconvolution
 def decon_lr(o_image, probe, tags, verbose=False):
     """
@@ -1258,795 +963,4 @@ def decon_lr(o_image, probe, tags, verbose=False):
     # plt.show()
     print(est2.shape)
     return est2
-
-
-##########################################
-# Functions Used
-##########################################
-
-
-def make_probe_gauss(size_x, size_y, width_i, xi, yi):
-    size_x = (size_x / 2)
-    size_y = (size_y / 2)
-    width = 2 * width_i ** 2
-    x, y = np.mgrid[-size_x:size_x, -size_y:size_y]
-    g = np.exp(-((x - xi) ** 2 / float(width) + (y - yi) ** 2 / float(width)))
-    probe = g / g.sum()
-
-    return probe
-
-
-def make_lorentz(size_x, size_y, width, xi, yi):
-    size_x = np.floor(size_x / 2)
-    size_y = np.floor(size_y / 2)
-    gamma = width
-    x, y = np.mgrid[-size_x:size_x, -size_y:size_y]
-    g = gamma / (2 * np.pi) / np.power(((x - xi) ** 2 + (y - yi) ** 2 + gamma ** 2), 1.5)
-    probe = g / g.sum()
-
-    return probe
-
-
-def ZLPWeight():
-    x = np.linspace(-0.5, 0.9, 29)
-    y = [0.0143, 0.0193, 0.0281, 0.0440, 0.0768, 0.1447, 0.2785, 0.4955, 0.7442, 0.9380,1.0000, 0.9483, 0.8596, 0.7620, 0.6539, 0.5515, 0.4478, 0.3500, 0.2683, 0.1979, 0.1410, 0.1021, 0.0752, 0.0545, 0.0401, 0.0300, 0.0229, 0.0176, 0.0139]
-    
-    return (x, y)
-
-
-    ##
-    # All atom detection is done here
-    # Everything is in unit of pixel!!
-    ##
-
-
-def findatoms(image, tags):
-    """
-    ######################################
-    # Find atoms 
-    ######################################
-    """
-    
-    image = image-image.min()
-    image = image/image.max()
-
-    if 'sigma_min' not in tags:
-        tags['sigma_min'] = 0.1
-    if 'resolution' not in tags:
-        tags['resolution'] = 0.1
-
-    if 'ROIsize' not in tags:
-        tags['ROIsize'] = 100.
-
-
-    res = tags['resolution'] / tags['pixel_size']#* tags['ROIsize']/100.
-    print('res',res)
-    coordinates = peak_local_max(image, min_distance=int(res/2), threshold_rel=tags['sigma_min'], exclude_border =True)
-    print('coor',len( coordinates))
-    """
-       peak_local_max(image, min_distance=10, threshold_abs=0, threshold_rel=0.1,
-                   exclude_border=True, indices=True, num_peaks=np.inf,
-                   footprint=None, labels=None):
-    
-        Find peaks in an image, and return them as coordinates or a boolean array.
-        Peaks are the local maxima in a region of `2 * min_distance + 1
-        
-        (i.e. peaks are separated by at least `min_distance`).
-        NOTE: If peaks are flat (i.e. multiple adjacent pixels have identical
-        intensities), the coordinates of all such pixels are returned.
-        """
-        
-
-
-
-    
-    # We calculate the radius in pixel of a round area in which atoms are evaluated
-    sc = tags['pixel_size']
-    r= tags['resolution']/sc*tags['ROIsize']/100./2.
-    tags['radius'] = r
-
-    #######################################
-    # Now we determine intensity #
-    #######################################
-
-    ###
-    # Make a circular mask for integration of atom intensity
-    ###
-    rr = int(r+0.5)
-    mask = np.zeros((2*rr+1,2*rr+1))
-    
-    for i in range (2*rr+1):
-        for j  in range (2*rr+1):
-            if (i-rr)**2+(j-rr)**2<rr**2+0.1:
-                mask[i,j]=1
-
-                
-                
-        
-                    
-    ###
-    # Determine  pixel position and intensity  of all atoms 
-    ###
-    atoms = []
-    for i in range(len( coordinates)):
-        x,y = coordinates[i]
-        
-        if x>rr and y>rr and x<image.shape[1]-rr and y<image.shape[0]-rr:
-            
-            area = image[x-rr:x+rr+1,y-rr:y+rr+1]
-            arr = area*mask
-            atoms.append((x,y,rr, arr.sum(), arr.max()))
-
-    print(' Detected ', len(atoms), ' atoms')
-    atoms.sort()
-    return atoms
-
-# sort corners in counter-clockwise direction
-def TurningFunction(corners,points):
-    # calculate centroid of the polygon
-    corners1 = np.array(points[corners])
-    corners2 = np.roll(corners1,1)
-    
-    corners0 = np.roll(corners1,-1)
-
-    v= corners1-corners0
-    an = (np.arctan2(v[:,0],v[:,1]) + 2.0 * np.pi)% (2.0 * np.pi)/np.pi*180
-    print(corners1)
-    print('an',an,v)
-    print(4*180/6)
-
-    angles = []
-    for i in range(len(corners1)):
-        A = corners1[i] - corners0[i]
-        B = corners1[i] - corners2[i] 
-        num = np.dot(A, B)
-        denom = np.linalg.norm(A) * np.linalg.norm(B)
-        angles.append(np.arccos(num/denom) * 180 / np.pi)
-        
-    return angles
-
-
-    
- 
-def PolygonSort2(corners,points):
-    """
-    # sort corners in counter-clockwise direction
-    input:
-            corners are indices in points array
-            points is list or array of points
-    output: 
-            cornersWithAngles
-    """
-    # calculate centroid of the polygon
-    n = len(corners) # of corners
-    cx = float(sum(x for x, y in points[corners])) / n
-    cy = float(sum(y for x, y in points[corners])) / n
-    
-    # create a new list of corners which includes angles
-    # angles from the positive x axis
-    cornersWithAngles = []
-    for i in corners:
-        x,y = points[i]       
-        an = (np.atan2(y - cy, x - cx) + 2.0 * np.pi)% (2.0 * np.pi)
-        cornersWithAngles.append([i, np.degrees(an)])
-    
-    #sort it using the angles
-    cornersWithAngles.sort(key = lambda tup: tup[1])
-    
-
-    return  cornersWithAngles
-
-
-def PolygonsInner(indices, points):
-    pp = np.array(points)[indices,:]
-    # Determine inner angle of polygon
-    # Generate second array which is shifted by one 
-    pp2 = np.roll(pp,1,axis=0)
-    # and subtract it from former: this is now a list of vectors
-    p_vectors = pp-pp2
-
-    #angles of vectors with respect to positive x-axis
-    ang = np.arctan2(p_vectors[:,1],p_vectors[:,0])/np.pi*180+360 % 360
-    # shift angles array by one  
-    ang2 = np.roll(ang,-1,axis=0)
-
-    #difference of angles is outer angle but we want the inner (inner + outer = 180) 
-    inner_angles = (180-(ang2-ang)+360 )% 360
-    
-    return inner_angles
-
-
- # sort corners in counter-clockwise direction
-def PolygonSort(corners):
-    # calculate centroid of the polygon
-    n = len(corners) # of corners
-    cx = float(sum(x for x, y in corners)) / n
-    cy = float(sum(y for x, y in corners)) / n
-    
-    # create a new list of corners which includes angles
-    cornersWithAngles = []
-    for x, y in corners:
-        an = (np.atan2(y - cy, x - cx) + 2.0 * np.pi)% (2.0 * np.pi)
-        cornersWithAngles.append((x, y, np.degrees(an)))
-    
-    #sort it using the angles
-    cornersWithAngles.sort(key = lambda tup: tup[2])
-
-    return  cornersWithAngles
-
-
-
-
-def PolygonArea(corners):
-    """
-    # Area of Polygon using Shoelace formula
-    # http://en.wikipedia.org/wiki/Shoelace_formula
-    # FB - 20120218
-    # corners must be ordered in clockwise or counter-clockwise direction
-    """
-    n = len(corners) # of corners
-    area = 0.0
-    C_x =0
-    C_y =0
-    for i in range(n):
-        j = (i + 1) % n
-        nn = corners[i][0] * corners[j][1] - corners[j][0] * corners[i][1]
-        area += nn
-        C_x +=  (corners[i][0] + corners[j][0])*nn
-        C_y +=  (corners[i][1] + corners[j][1])*nn
-
-    area = abs(area) / 2.0
-
-    # centeroid or arithmetic mean 
-    C_x = C_x/(6*area) 
-    C_y = C_y/(6* area)
-
-    return (area), C_x, C_y
-
-
-def PolygonAngles( corners):
-    angles = []
-    # calculate centroid of the polygon
-    n = len(corners) # of corners
-    cx = float(sum(x for x, y in corners)) / n
-    cy = float(sum(y for x, y in corners)) / n
-    # create a new list of angles
-    #print (cx,cy)
-    for x, y in corners:
-        an = (np.atan2(y - cy, x - cx) + 2.0 * np.pi)% (2.0 * np.pi)
-        angles.append((np.degrees(an)))
-
-    return angles
-
-
-
-
-
-
-def voronoi_tags(vor):
-    sym = {}
-    sym['voronoi'] = vor
-    sym['vertices'] = vor.vertices #(ndarray of double, shape (nvertices, ndim)) Coordinates of the Voronoi vertices.
-    sym['ridge_points'] = vor.ridge_points #ridge_points 	(ndarray of ints, shape (nridges, 2)) Indices of the points between which each Voronoi ridge lies.
-    sym['ridge_vertices'] = vor.ridge_vertices #ridge_vertices 	(list of list of ints, shape (nridges, *)) Indices of the Voronoi vertices forming each Voronoi ridge.
-    sym['regions'] = vor.regions #regions 	(list of list of ints, shape (nregions, *)) Indices of the Voronoi vertices forming each Voronoi region. -1 indicates vertex outside the Voronoi diagram.
-    sym['point_region'] = vor.point_region #point_region 	(list of ints, shape (npoints)) Index of the Voronoi region for each input point. If qhull option “Qc” was not specified, the list will contain -1 for points that are not associated with a Voronoi region.
-
-    
-    points = vor.points
-    nnTree = KDTree(points)
-    
-    rim = []
-    regions=[]
-
-    ###
-    # We get all the vertice length
-  
-    lengths =[]
-    for vertice in  vor.ridge_vertices:
-        if not(-1 in vertice):
-            p1 = vor.vertices[vertice[0]]
-            p2 = vor.vertices[vertice[1]]
-            lengths.append(np.sqrt((p1[0]-p2[0])**2+(p1[1]-p2[1] )**2))
-                
-    sym['lengths'] = lengths
-    sym['median lengths'] = np.median(lengths)
-    sym['Min Voronoi Edge'] = np.median(lengths)/1.5
-    #print ('median lengths', np.median(lengths))
-    #print ('Min Voronoi Edge',np.median(lengths)/1.5)
-    cornersHist=[]
-    nnHist = []
-    nnDistHist =[]
-    angleHist = []
-    areaHist=[]
-    deviationHist =[]
-
-    for i, region in enumerate(vor.point_region):
-        x,y = points[i]
-        sym[str(i)]={}
-        vertices = vor.regions[region]
-
-        ###
-        # We get all the rim atoms
-        ###
-
-        #if all(v >= 0  and all(vor.vertices[v] >0) and all(vor.vertices[v]<tags['data'].shape[0]) for v in vertices):
-        if all(v >= 0  and all(vor.vertices[v] >0) for v in vertices):
-             # finite regions only now
-            # negative and too large vertices (corners) are excluded
-
-            regions.append(vertices)
-            poly = []
-            for v in vertices:
-                poly.append(vor.vertices[v])
-
-            area, cx,cy = PolygonArea(poly)
-            cx = abs(cx)
-            cy = abs(cy)
-
-            angles = PolygonAngles(poly)
-            angleHist.append(angles)
-            areaHist.append(area)
-            deviationHist.append(np.sqrt((cx-x)**2+ (cy-y)**2))
-
-            sym[str(i)]['xy'] = [x, y]
-            sym[str(i)]['geometric'] = [cx, cy]
-            sym[str(i)]['area'] = area
-
-            sym[str(i)]['angles'] = angles
-            sym[str(i)]['off center'] = [cx-x, cy-y]
-
-            sym[str(i)]['position'] = 'inside'
-            sym[str(i)]['corner'] = vertices
-            sym[str(i)]['vertices']=poly
-            sym[str(i)]['corners'] = len(vertices)
-            cornersHist.append(len(vertices))
-            nn = 0
-            nnVor = []
-            length = []
-            for j in range(len(vertices)):
-                k = (j+1) % len(vertices)
-                p1 = vor.vertices[vertices[j]]
-                p2 = vor.vertices[vertices[k]]
-                leng = np.sqrt((p1[0]-p2[0])**2+(p1[1]-p2[1] )**2)
-                length.append(leng)
-                sym[str(i)]['length'] = length
-                if leng > sym['Min Voronoi Edge']:        
-                    nn +=1
-                    nnVor.append(vertices[j])
-                sym[str(i)]['length'] = length    
-            nnP = nnTree.query(points[i],k = nn+1)
-            sym [str(i)]['neighbors'] = []
-            sym [str(i)]['nn Distance'] = []
-            sym [str(i)]['nn']=nn
-            if nn>0:
-                nnHist.append(nn)   
-                for j in range (1,len(nnP[0])):
-                    sym [str(i)]['nn Distance'].append(nnP[0][j])
-                    sym [str(i)]['neighbors'].append(nnP[1][j])
-                    nnDistHist.append(nnP[0][j])
-            else:
-                rim.append(i)
-                sym[str(i)]['position'] = 'rim'
-                sym[str(i)]['corners'] = 0
-                print('weird nn determination',i)
-
-        else:
-            rim.append(i)
-            sym[str(i)]['position'] = 'rim'
-            sym[str(i)]['corners'] = 0
-            sym[str(i)]['xy'] = [x, y]
-
-
-    sym['average corners']= np.median(cornersHist)
-    sym['average area']= np.median(areaHist)
-    sym['num atoms at rim']= len(rim)
-    sym['num voronoi']= len(points)-len(rim)
-    sym['Median Coordination']= np.median(nnHist)
-    sym['Median NN Distance']= np.median(nnDistHist)
-
-    sym['Hist corners']= (cornersHist)
-    sym['Hist area']= areaHist
-    sym['atoms at rim']= (rim)
-    sym['Hist Coordination']= (nnHist)
-    sym['Hist NN Distance']= (nnDistHist)
-    sym['Hist deviation']= (deviationHist)
-
-
-    return sym
-    #print ('average corners', np.median(cornersHist))
-
-
-def defineSymmetry(tags):
-
-    #make dictionary to store
-    if 'symmetry' in tags:
-        tags['symmetry'].clear()
-
-    tags['symmetry'] = {}
-    sym = tags['symmetry']
-    if 'latticeType' in tags:
-        latticeTypes = ['None', 'Find Lattice',  'hexagonal', 'honeycomb', 'square', 'square centered',
-                'diamond', 'fcc']
-        sym['lattice']=latticeTypes[tags['latticeType']]
-
-    sym['number of atoms'] = len(tags['atoms'])
-    
-    points = []
-    for i in range(sym['number of atoms']):            
-        sym[str(i)] = {}
-        sym[str(i)]['index']= i
-        sym[str(i)]['x'] = tags['atoms'][i][0]
-        sym[str(i)]['y'] = tags['atoms'][i][1]
-        sym[str(i)]['intensity'] = tags['atoms'][i][3]
-        sym[str(i)]['maximum'] = tags['atoms'][i][4]
-        sym[str(i)]['position'] = 'inside'
-        sym[str(i)]['Z'] = 0
-        sym[str(i)]['Name'] = 'undefined'
-        sym[str(i)]['Column'] = -1
-        
-        points.append([int(sym[str(i)]['x']+0.5),int(sym[str(i)]['y']+0.5)])
-
-    points = points.copy()
-    
-
-
-def voronoi2(tags, atoms):
-
-    sym = tags['symmetry']
-    points = []
-    
-    for i in range(sym['number of atoms']):            
-        points.append([int(sym[str(i)]['x']+0.5),int(sym[str(i)]['y']+0.5)])
-
-
-    #points = np.array(atoms[:][0:2])
-    vor = sp.Voronoi(points)
-
-    
-    sym['voronoi'] = vor
-
-    nnTree = sp.KDTree(points)
-    
-    rim = []
-    regions=[]
-
-    ###
-    # We get all the vertice length
-  
-    lengths =[]
-    for vertice in  vor.ridge_vertices:
-        if all(v >= 0 for v in vertice):
-            p1 = vor.vertices[vertice[0]]
-            p2 = vor.vertices[vertice[1]]
-            lengths.append(np.sqrt((p1[0]-p2[0])**2+(p1[1]-p2[1] )**2))
-                
-    sym['lengths'] = lengths
-    sym['median lengths'] = np.median(lengths)
-    sym['Min Voronoi Edge'] = np.median(lengths)/1.5
-    #print ('median lengths', np.median(lengths))
-    #print ('Min Voronoi Edge',np.median(lengths)/1.5)
-    cornersHist=[]
-    nnHist = []
-    nnDistHist =[]
-    angleHist = []
-    areaHist=[]
-    deviationHist =[]
-
-    for i, region in enumerate(vor.point_region):
-        x,y = points[i]
-
-        vertices = vor.regions[region]
-
-        ###
-        # We get all the rim atoms
-        ###
-
-        if all(v >= 0  and all(vor.vertices[v] >0) and all(vor.vertices[v]<tags['data'].shape[0]) for v in vertices):
-            # finite regions only now
-            # negative and too large vertices (corners) are excluded
-
-            regions.append(vertices)
-            poly = []
-            for v in vertices:
-                poly.append(vor.vertices[v])
-
-            area, cx,cy = PolygonArea(poly)
-            cx = abs(cx)
-            cy = abs(cy)
-
-            angles = PolygonAngles(poly)
-            angleHist.append(angles)
-            areaHist.append(area)
-            deviationHist.append(np.sqrt((cx-x)**2+ (cy-y)**2))
-
-            sym[str(i)]['xy'] = [x, y]
-            sym[str(i)]['geometric'] = [cx, cy]
-            sym[str(i)]['area'] = area
-
-            sym[str(i)]['angles'] = angles
-            sym[str(i)]['off center'] = [cx-x, cy-y]
-
-            sym[str(i)]['position'] = 'inside'
-            sym[str(i)]['corner'] = vertices
-            sym[str(i)]['vertices']=poly
-            sym[str(i)]['corners'] = len(vertices)
-            cornersHist.append(len(vertices))
-            nn = 0
-            nnVor = []
-            length = []
-            for j in range(len(vertices)):
-                k = (j+1) % len(vertices)
-                p1 = vor.vertices[vertices[j]]
-                p2 = vor.vertices[vertices[k]]
-                leng = np.sqrt((p1[0]-p2[0])**2+(p1[1]-p2[1] )**2)
-                length.append(leng)
-                sym[str(i)]['length'] = length
-                if leng > sym['Min Voronoi Edge']:        
-                    nn +=1
-                    nnVor.append(vertices[j])
-                sym[str(i)]['length'] = length    
-            nnP = nnTree.query(points[i],k = nn+1)
-            sym [str(i)]['neighbors'] = []
-            sym [str(i)]['nn Distance'] = []
-            sym [str(i)]['nn']=nn
-            if nn>0:
-                nnHist.append(nn)   
-                for j in range (1,len(nnP[0])):
-                    sym [str(i)]['nn Distance'].append(nnP[0][j])
-                    sym [str(i)]['neighbors'].append(nnP[1][j])
-                    nnDistHist.append(nnP[0][j])
-            else:
-                rim.append(i)
-                sym[str(i)]['position'] = 'rim'
-                sym[str(i)]['corners'] = 0
-                print('weird nn determination',i)
-
-        else:
-            rim.append(i)
-            sym[str(i)]['position'] = 'rim'
-            sym[str(i)]['corners'] = 0
-            sym[str(i)]['xy'] = [x, y]
-
-
-    sym['average corners']= np.median(cornersHist)
-    sym['average area']= np.median(areaHist)
-    sym['num atoms at rim']= len(rim)
-    sym['num voronoi']= len(points)-len(rim)
-    sym['Median Coordination']= np.median(nnHist)
-    sym['Median NN Distance']= np.median(nnDistHist)
-
-    sym['Hist corners']= (cornersHist)
-    sym['Hist area']= areaHist
-    sym['atoms at rim']= (rim)
-    sym['Hist Coordination']= (nnHist)
-    sym['Hist NN Distance']= (nnDistHist)
-    sym['Hist deviation']= (deviationHist)
-
-
-
-    #print ('average corners', np.median(cornersHist))
-
-def intensity_area(image,atoms, radius):
-    rr = int(radius+0.5) # atom radius
-    print('using radius ',rr, 'pixels')
-    
-    pixels = np.linspace(0,2*rr,2*rr+1)-rr
-    x,y = np.meshgrid(pixels,pixels);
-    mask = (x**2+y**2) < rr**2 #
-    intensity_area = []
-    for i in range(len( atoms)):
-        
-        x = int(atoms[i][1]   ) 
-        y = int(atoms[i][0]   ) 
-        area = image[x-rr:x+rr+1,y-rr:y+rr+1]
-        if area.shape == mask.shape:
-            intensity_area.append((area*mask).sum() )
-        else:
-            intensity_area.append(-1)
-    return intensity_area
-
-def Gauss_2D(params, ydata):
-    width = int(ydata.shape[0]/2)
-    Gauss_width = params[0]
-    x0 = params[1]
-    y0 = params[2]
-    inten = params[3]
-
-    x, y = np.mgrid[-width:width+1, -width:width+1]
-
-
-    return np.exp(-((x-x0)**2 + (y-y0)**2) /2./ Gauss_width**2)*inten
-def Gauss_difference (params,  xdata, ydata):
-    #self.img1b.setImage(gauss)
-    gauss = Gauss_2D(params, ydata)
-    return (ydata - gauss).flatten()
-
-def atomRefine(image, atoms, radius, MaxInt = 0,MinInt = 0, maxDist = 4):
-    """
-        fits a Gaussian in a blob
-    """
-    rr = int(radius+0.5) # atom radius
-    print('using radius ',rr, 'pixels')
-    
-    pixels = np.linspace(0,2*rr,2*rr+1)-rr
-    x,y = np.meshgrid(pixels,pixels);
-    mask = (x**2+y**2) < rr**2 #
-
-    guess  = [rr*2, 0.0, 0.0 , 1]    
-    
-    sym = {}
-    sym['number_of_atoms'] = len(atoms)
-    
-    volume = []
-    position = []
-    intensity_area = []
-    maximum_area = []
-    newatoms = []
-    Gauss_width = []
-    Gauss_amplitude = []
-    Gauss_intensity = []
-    
-    done = 0
-    for i in range(len(atoms)):
-        if done < int((i+1)/len(atoms)*50):
-            done = int((i+1)/len(atoms)*50)
-            sys.stdout.write('\r')
-            # progress output :
-            sys.stdout.write("[%-50s] %d%%" % ('='*done, 2*done))
-            sys.stdout.flush()
-        
-        y,x = atoms[i][0:2]
-        x = int(x)
-        y = int(y)
-        append = False
-        
-        
-        area = image[x-rr:x+rr+1,y-rr:y+rr+1]
-       
-        append = False
-        
-        if (x-rr) < 0 or y-rr <0 or x+rr+1 > image.shape[0] or y+rr+1 > image.shape[1]:
-            position.append(-1)
-            intensity_area.append(0) 
-            maximum_area.append(0)
-        else:
-            position.append(1)
-            intensity_area.append((area*mask).sum() )
-            maximum_area.append((area*mask).max())
-            
-        if MaxInt>0:
-            if area.sum()< MaxInt:                    
-                if area.sum() > MinInt:
-                    append = True
-        elif area.sum()> MinInt:
-            append = True
-        
-        pout = [0,0,0,0]
-        if append:
-            if (x-rr) < 0 or y-rr <0 or x+rr+1 > image.shape[0] or y+rr+1 > image.shape[1]:
-                pass
-            else:
-                pout, res =  leastsq(Gauss_difference, guess, args=(area, area))
-                
-            if (abs(pout[1])> maxDist) or (abs(pout[2])> maxDist):
-                pout = [0,0,0,0]
-    
-        volume.append(2* np.pi * pout[3] * pout[0]*pout[0])
-
-        newatoms.append([y+pout[2], x+pout[1]])# ,pout[0],  volume)) #,pout[3]))
-        if (all(v == 0 for v in pout)):
-            Gauss_intensity.append(0.)
-        else:
-            Gauss_intensity.append((Gauss_2D(pout, area)*mask).sum() )
-        Gauss_width.append(pout[0])
-        Gauss_amplitude.append(pout[3])
-    
-    
-    sym['inside'] = position
-    sym['intensity_area'] = intensity_area 
-    sym['maximum_area'] = maximum_area
-    sym['atoms'] = newatoms
-    sym['Gauss_width'] = Gauss_width
-    sym['Gauss_amplitude'] = Gauss_amplitude
-    sym['Gauss_intensity'] = Gauss_intensity
-    sym['Gauss_volume'] = volume
-    
-    return sym
-
-def Fourier_transform(current_channel,data):# = image_channel
-    # spatial data
-    tags = dict(current_channel.attrs)
-    out_tags = {}
-    basename = current_channel['title'][()]
-    
-    sizeX = current_channel['spatial_size_x'][()]
-    sizeY = current_channel['spatial_size_y'][()]
-    scaleX = current_channel['spatial_scale_x'][()]
-    scaleY = current_channel['spatial_scale_y'][()]
-    basename = current_channel['title'][()]
-
-    FOV_x = sizeX*scaleX
-    FOV_y = sizeY*scaleY
-    
-    image = data- data.min()
-    fft_mag = (np.abs((np.fft.fftshift(np.fft.fft2(image)))))
-    
-    out_tags['Magnitude']=fft_mag
-
-    ## pixel_size in recipical space
-    rec_scale_x = 1/FOV_x  
-    rec_scale_y = 1/FOV_y 
-
-    ## Field of View (FOV) in recipical space please note: rec_FOV_x = 1/(scaleX*2)
-    rec_FOV_x = rec_scale_x * sizeX /2.
-    rec_FOV_y = rec_scale_y * sizeY /2.
-    print(rec_FOV_x , 1/(scaleX*2))
-
-
-    ## Field ofView (FOV) in recipical space
-    rec_extend = (-rec_FOV_x,rec_FOV_x-rec_scale_x,rec_FOV_y-rec_scale_y,-rec_FOV_y)
-
-    out_tags['spatial_size_x']=sizeX
-    out_tags['spatial_size_y']=sizeY
-    out_tags['spatial_scale_x']=rec_scale_x
-    out_tags['spatial_scale_y']=rec_scale_y
-    out_tags['spatial_origin_x']=sizeX/2.
-    out_tags['spatial_origin_y']=sizeY/2.
-    out_tags['title']=out_tags['basename']=basename
-    out_tags['FOV_x']=rec_FOV_x
-    out_tags['FOV_y']=rec_FOV_y
-    out_tags['extent']=rec_extend
-    
-    
-    # We need some smoothing (here with a Gaussian)
-    smoothing = 3
-    fft_mag2 = ndimage.gaussian_filter(fft_mag, sigma=(smoothing, smoothing), order=0)
-    #fft_mag2 = np.log2(1+fft_mag2)
-
-    out_tags['data'] = out_tags['Magnitude_smoothed']=fft_mag2
-    #prepare mask
-    pixelsy = (np.linspace(0,image.shape[0]-1,image.shape[0])-image.shape[0]/2)* rec_scale_x
-    pixelsx = (np.linspace(0,image.shape[1]-1,image.shape[1])-image.shape[1]/2)* rec_scale_y
-    x,y = np.meshgrid(pixelsx,pixelsy);
-    mask = np.zeros(image.shape)
-
-    mask_spot = x**2+y**2 > 1**2 
-    mask = mask + mask_spot
-    mask_spot = x**2+y**2 < 11**2 
-    mask = mask + mask_spot
-
-    mask[np.where(mask==1)]=0 # just in case of overlapping disks
-
-    minimum_intensity = np.log2(1+fft_mag2)[np.where(mask==2)].min()*0.95
-    #minimum_intensity = np.mean(fft_mag3)-np.std(fft_mag3)
-    maximum_intensity = np.log2(1+fft_mag2)[np.where(mask==2)].max()*1.05
-    #maximum_intensity =  np.mean(fft_mag3)+np.std(fft_mag3)*2
-    out_tags['minimum_intensity']=minimum_intensity
-    out_tags['maximum_intensity']=maximum_intensity
-    
-    return out_tags
-
-
-
-def find_Bragg(fft_tags, spot_threshold = 0 , verbose = False):
-    if spot_threshold ==0:
-        spot_threshold = 0.05#(fft_tags['maximum_intensity']*10)
-    
-    # we'll have to switch x and ycoordonates
-    center = np.array([int(fft_tags['spatial_origin_y']), int(fft_tags['spatial_origin_x']),1] )
-    rec_scale = np.array([fft_tags['spatial_scale_y'], fft_tags['spatial_scale_x'],1])
-    data = fft_tags['data'].T
-    data = (data-data.min())/data.max()
-    spots_random =  (blob_log(data,  max_sigma= 5 , threshold=spot_threshold)-center)*rec_scale
-    
-    if verbose:
-        print(f'found {len(spots_random)} Bragg spots with threshold of {spot_threshold}')
-    spots_random[:,2] = np.linalg.norm(spots_random[:,0:2], axis=1)
-    spots_index = np.argsort(spots_random[:,2])
-    
-    spots = spots_random[spots_index]
-    spots[:,2] = np.arctan2(spots[:,0], spots[:,1])
-    return spots
-
